오픈 마켓 [STEP Bonus] 아리, 제리

+++ OpenMarket/OpenMarket/Model/Response/Product.swift
@@ -13,9 +13,9 @@ struct Product: Codable {
     let name: String
     var thumbnail: String
     var currency: Currency
-    var price: Int
-    var bargainPrice: Int
-    var discountedPrice: Int
+    var price: Double
+    var bargainPrice: Double
+    var discountedPrice: Double
     var stock: Int
     var createdAt: String
     var issuedAt: String
